Cargo: Update dependencies.

+++ source/vm/engine.rs
@@ -128,7 +128,7 @@ impl Engine {
                 &mut engine_ref,
                 module.to_ref(),
                 &mut engine_options,
-                engine_options_size as usize,
+                engine_options_size as u64,
                 &mut engine_error
             );
         }
